Transparent Window Update
Removed all customization capabilities. Only works for Windows on CUDA
Following that, the method of display is no longer by typing directly. It now has a transparent floating window that is anchored to the left-hand side of the mouse
The window is always on top and follows the mouse around and the keyboard shortcuts to enact the transcription is control shift alt to record and stop recording
Following that the input is no longer simulated onto the current cursor field however it is simulated typing onto the transparent window to encourage reading of the transcription
However, if the user so pleases, they are able to instantly press control V to paste and not have to wait to see what is paste simulator typed in the floating window
Control shift alt enacts the recording
Control-Shift-Alt also stops the recording, at which point the text is simulated typed out onto the transparent window
and control v paste the text because as a backup slash the only real way it stores text other than the floating window is within the clipboard